change naming convention and move interfaces to place where they are needed
